feat(DTFS2-7674): add submitDealToTfm test helper
