+++ dtfs-central-api/api-tests/v1/portal-facility-amendments/amendment-put.api-test.ts
@@ -1,21 +1,11 @@
 import { Response } from 'supertest';
 import { ObjectId } from 'mongodb';
 import { HttpStatusCode } from 'axios';
-import {
-  AnyObject,
-  API_ERROR_CODE,
-  DEAL_SUBMISSION_TYPE,
-  DEAL_TYPE,
-  FACILITY_TYPE,
-  MONGO_DB_COLLECTIONS,
-  PortalFacilityAmendment,
-  TFM_DEAL_STAGE,
-} from '@ukef/dtfs2-common';
+import { AnyObject, API_ERROR_CODE, DEAL_SUBMISSION_TYPE, DEAL_TYPE, FACILITY_TYPE, MONGO_DB_COLLECTIONS, PortalFacilityAmendment } from '@ukef/dtfs2-common';
 import { generatePortalAuditDetails } from '@ukef/dtfs2-common/change-stream';
 import wipeDB from '../../wipeDB';
 import { testApi } from '../../test-api';
-import { createDeal } from '../../helpers/create-deal';
-import { MOCK_PORTAL_USER } from '../../mocks/test-users/mock-portal-user';
+import { createDeal, submitDealToTfm } from '../../helpers/create-deal';
 import aDeal from '../deal-builder';
 import { aPortalUser } from '../../mocks/test-users/portal-user';
 import { createPortalUser } from '../../helpers/create-portal-user';
@@ -57,34 +47,7 @@ describe('PUT /v1/portal/facilities/:facilityId/amendments/', () => {
 
     facilityId = createFacilityResponse.body._id;
 
-    await testApi
-      .put({
-        dealType: DEAL_TYPE.GEF,
-        dealId,
-        submissionType: DEAL_SUBMISSION_TYPE.AIN,
-        auditDetails: generatePortalAuditDetails(MOCK_PORTAL_USER._id),
-      })
-      .to('/v1/tfm/deals/submit');
-
-    await testApi
-      .put({
-        dealUpdate: {
-          tfm: {
-            dateReceived: '23-09-2024',
-            dateReceivedTimestamp: 1727085149,
-            parties: {},
-            activities: [],
-            product: DEAL_TYPE.GEF,
-            stage: TFM_DEAL_STAGE.CONFIRMED,
-            exporterCreditRating: 'Acceptable (B+)',
-            lastUpdated: 1727085149571,
-            lossGivenDefault: 50,
-            probabilityOfDefault: 12,
-          },
-        },
-        auditDetails: generatePortalAuditDetails(MOCK_PORTAL_USER._id),
-      })
-      .to(`/v1/tfm/deals/${dealId}`);
+    await submitDealToTfm({ dealId, dealSubmissionType: DEAL_SUBMISSION_TYPE.AIN, dealType: DEAL_TYPE.GEF });
   });
 
   afterAll(() => {
